Completed the authentication service
