--- authentication_service/src/main/java/com/eventu/authentication_service/controllers/response/AuthResponse.java
@@ -1,13 +1,0 @@
-package com.eventu.authentication_service.controllers.response;
-
-import lombok.Data;
-
-@Data
-public class AuthResponse {
-
-    private String jwt;
-
-    private String message;
-
-    private Boolean status;
-}
